+++ apps/portal/controllers/apis/database.jag
@@ -1,0 +1,39 @@
+<%
+/*
+ * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+(function () {
+    if (!user) {
+        response.sendError(401, 'unauthorized to access assets');
+        return;
+    }
+
+    var log = new Log();
+    var dataBaseUtils = require('/modules/database-utils.js');
+    var task = request.getParameter('task');
+    var INSERT_TASK = 'insert';
+    var i18n = require('i18n');
+    i18n.init(request, utils.getLocaleResourcePath());
+
+    if (request.getMethod('POST')) {
+        if (task === insert) {
+
+        }
+    } else {
+        response.sendError(403,  i18n.localize("forbidden"));
+    }
+}());
+
+%>
