+++ features/org.wso2.carbon.dashboards.portal.feature/src/main/resources/dashboards/h2.sql
@@ -22,4 +22,4 @@ CREATE TABLE IF NOT EXISTS GADGET_USAGE (
     USAGE_DATA TEXT,
     PRIMARY KEY(TENANT_ID, DASHBOARD_ID, GADGET_ID)
 );
-CREATE INDEX G_INDEX ON GADGET_USAGE(TENANT_ID, DASHBOARD_ID, GADGET_ID);
+CREATE INDEX IF NOT EXISTS G_INDEX ON GADGET_USAGE(TENANT_ID, DASHBOARD_ID, GADGET_ID);
