+++ features/org.wso2.carbon.dashboards.portal.feature/src/main/resources/dashboards/mysql5.7.sql
@@ -21,4 +21,4 @@ CREATE TABLE IF NOT EXISTS GADGET_USAGE (
     USAGE_DATA TEXT,
     PRIMARY KEY(TENANT_ID, DASHBOARD_ID, GADGET_ID)
 )ENGINE INNODB;
-CREATE INDEX G_INDEX ON GADGET_USAGE(TENANT_ID, DASHBOARD_ID, GADGET_ID);
+CREATE INDEX G_INDEX USING HASH ON GADGET_USAGE(TENANT_ID, DASHBOARD_ID, GADGET_ID);
