+++ features/org.wso2.carbon.dashboards.portal.feature/src/main/resources/p2.inf
@@ -12,5 +12,4 @@ org.eclipse.equinox.p2.touchpoint.natives.copy(source:${installFolder}/../featur
 org.eclipse.equinox.p2.touchpoint.natives.copy(source:${installFolder}/../features/org.wso2.carbon.dashboards.portal_${feature.version}/portal/store/carbon.super/fs/layout,target:${installFolder}/../../resources/sharedstore/layout,overwrite:true);\
 org.eclipse.equinox.p2.touchpoint.natives.copy(source:${installFolder}/../features/org.wso2.carbon.dashboards.portal_${feature.version}/portal.json,target:${installFolder}/../../conf/dashboards/portal.json,overwrite:true);\
 org.eclipse.equinox.p2.touchpoint.natives.copy(source:${installFolder}/../features/org.wso2.carbon.dashboards.portal_${feature.version}/dashboards,target:${installFolder}/../../../dbscripts/dashboards,overwrite:true);\
-org.eclipse.equinox.p2.touchpoint.natives.copy(source:${installFolder}/../features/org.wso2.carbon.dashboards.portal_${feature.version}/WSO2DS_DB.h2.db/,target:${installFolder}/../../database/WSO2DS_DB.h2.db,overwrite:true);\
 org.eclipse.equinox.p2.touchpoint.natives.copy(source:${installFolder}/../features/org.wso2.carbon.dashboards.portal_${feature.version}/ds-datasources.xml,target:${installFolder}/../../conf/datasources/ds-datasources.xml,overwrite:true);\
